+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2026-03-02 04:00:33
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 6cc8f196407e2c0c47c78f685c0cbba6f17b8037
+
+## [ForwardWidgets-2kuai] - 2026-03-02 04:00:33
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: b6154c5434d62d97134ad80287a2ca96618cb0b9
+
+## [ForwardWidgets-pack1r] - 2026-03-02 04:00:33
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: d36b7c0558abdf7e13dcc5880b1c3b42309efaeb
+
+
 ## [ForwardWidgets-huangxd] - 2026-03-01 04:09:35
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main